pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tree-models"
version = "1.0.0"
description = "A comprehensive machine learning framework for tree-based models with advanced configuration management"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Aniket Sharma", email = "aniket.sharma@example.com"}
]
maintainers = [
    {name = "ML Team", email = "ml-team@example.com"}
]
keywords = [
    "machine learning", "xgboost", "lightgbm", "catboost", 
    "mlops", "fraud detection", "feature selection", "hyperparameter tuning"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "xgboost>=1.6.0",
    "lightgbm>=3.3.0",
    "catboost>=1.0.0",
    "optuna>=3.0.0",
    "mlflow>=2.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0", 
    "shap>=0.41.0",
    "joblib>=1.1.0",
    "tqdm>=4.62.0",
    "python-dateutil>=2.8.0",
    "PyYAML>=6.0",
    "pydantic>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0", 
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "flake8-docstrings>=1.6.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
]
database = [
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "pymongo>=4.0.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=8.5.0",
    "mkdocstrings>=0.19.0",
    "mkdocstrings-python>=0.7.0",
]
all = [
    "tree-models[dev,database,docs]"
]

[project.urls]
Homepage = "https://github.com/aniketsharma21/tree_models"
Documentation = "https://tree-models.readthedocs.io/"
Repository = "https://github.com/aniketsharma21/tree_models.git"
"Bug Tracker" = "https://github.com/aniketsharma21/tree_models/issues"
Changelog = "https://github.com/aniketsharma21/tree_models/blob/main/CHANGELOG.md"

[project.scripts]
tree-models = "tree_models.cli:main"

[project.entry-points."tree_models.plugins"]
xgboost = "tree_models.plugins.xgboost:XGBoostPlugin"
lightgbm = "tree_models.plugins.lightgbm:LightGBMPlugin"
catboost = "tree_models.plugins.catboost:CatBoostPlugin"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["tree_models*", "config*", "src*"]
exclude = ["tests*", "docs*", "notebooks*"]

[tool.setuptools.package-data]
"config" = ["defaults/*.yaml", "*.md"]
"tree_models" = ["py.typed"]

# Testing configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra", 
    "--strict-markers",
    "--strict-config",
    "--cov=tree_models",
    "--cov=config", 
    "--cov=src",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=80"
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "config: marks tests related to configuration system",
    "models: marks tests related to model training/evaluation",
    "data: marks tests related to data processing",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["tree_models", "config", "src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
    "*/setup.py",
    "*/conftest.py",
    "*/notebooks/*",
]
parallel = true
branch = true

[tool.coverage.paths]
source = [
    "tree_models/",
    "config/",
    "src/",
    "*/site-packages/tree_models/",
    "*/site-packages/config/",
    "*/site-packages/src/",
]

[tool.coverage.report]
show_missing = true
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Code formatting
[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = "\\.pyi?$"
extend-exclude = """
/(
    \\.eggs
  | \\.git
  | \\.hg
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
"""

# Import sorting
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["tree_models", "config", "src"]
known_third_party = [
    "pandas", "numpy", "sklearn", "xgboost", "lightgbm", 
    "catboost", "optuna", "mlflow", "matplotlib", "seaborn", "shap"
]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Type checking
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_error_context = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "xgboost.*",
    "lightgbm.*", 
    "catboost.*",
    "optuna.*",
    "mlflow.*",
    "shap.*",
    "boruta.*"
]
ignore_missing_imports = true

# Linting
[tool.flake8]
max-line-length = 100
extend-ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long
    "W503",  # line break before binary operator
]
exclude = [
    ".git",
    "__pycache__",
    "docs/",
    "build/",
    "dist/",
    "*.egg-info/",
    ".venv/",
    ".mypy_cache/",
    ".pytest_cache/",
]
per-file-ignores = [
    "__init__.py:F401",  # imported but unused
    "tests/*:D100,D101,D102,D103,D104",  # missing docstrings in tests
]

# Security linting
[tool.bandit]
exclude_dirs = ["tests", "build", "dist"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection (paramiko)

# Documentation
[tool.mkdocs]
site_name = "Tree Models Documentation"
site_description = "Comprehensive ML framework for tree-based models"
site_url = "https://tree-models.readthedocs.io/"
repo_url = "https://github.com/aniketsharma21/tree_models"
edit_uri = "edit/main/docs/"

[build]
# Build configuration
skip_build = false